etl/requirements.txt
peewee==3.13.1
tqdm==4.24.0
marshmallow==3.3.0
mysql-connector==2.2.9
PyMySQL==0.9.3
